+++ data/tasks.txt
@@ -1,0 +1,1 @@
+D | 1 | 0 | return book | 2/12/2025 1800
