--- src/main/java/Command.java
@@ -1,6 +1,0 @@
-import java.util.ArrayList;
-
-abstract class Command {
-
-    abstract boolean execute(TaskList list, Storage storage, UI ui);
-}
