--- src/main/java/ExitCommand.java
@@ -1,10 +1,0 @@
-class ExitCommand extends Command {
-
-    ExitCommand() {}
-
-    @Override
-    boolean execute(TaskList list, Storage storage, UI ui) {
-        ui.end();
-        return true;
-    }
-}
