--- src/main/java/ListCommand.java
@@ -1,10 +1,0 @@
-class ListCommand extends Command {
-
-    ListCommand() {}
-
-    @Override
-    boolean execute(TaskList list, Storage storage, UI ui) {
-        list.listTasks(ui);
-        return false;
-    }
-}
