--- src/main/java/NextCommand.java
@@ -1,11 +1,0 @@
-//dummy command
-
-class NextCommand extends Command {
-
-    NextCommand() {}
-
-    @Override
-    boolean execute(TaskList list, Storage storage, UI ui) {
-        return false;
-    }
-}
