+++ src/main/java/talkgpt/TalkGPTException.java
@@ -1,3 +1,5 @@
+package talkgpt;
+
 public class TalkGPTException extends Exception {
     public TalkGPTException(String message) {
         super(message);
