+++ src/main/java/talkgpt/command/Command.java
@@ -1,0 +1,10 @@
+package talkgpt.command;
+
+import talkgpt.ui.UI;
+import talkgpt.TaskList;
+import talkgpt.storage.Storage;
+
+public abstract class Command {
+
+    abstract public boolean execute(TaskList list, Storage storage, UI ui);
+}
