+++ src/main/java/talkgpt/command/DeadlineCommand.java
@@ -1,16 +1,23 @@
-class DeadlineCommand extends Command {
+package talkgpt.command;
+
+import talkgpt.ui.UI;
+import talkgpt.TaskList;
+import talkgpt.storage.Storage;
+import talkgpt.task.*;
+
+public class DeadlineCommand extends Command {
 
     private String description;
     private String deadline;
     private static final int INDEX_OFFSET = 1;
 
-    DeadlineCommand(String description, String deadline) {
+    public DeadlineCommand(String description, String deadline) {
         this.description = description;
         this.deadline = deadline;
     }
 
     @Override
-    boolean execute(TaskList list, Storage storage, UI ui) {
+    public boolean execute(TaskList list, Storage storage, UI ui) {
         Task newTask = new Deadline(list.size() + INDEX_OFFSET, description, deadline);
         list.addTask(newTask, storage, ui);
         return false;
