+++ src/main/java/talkgpt/command/ListCommand.java
@@ -1,0 +1,17 @@
+package talkgpt.command;
+
+import talkgpt.ui.UI;
+import talkgpt.TaskList;
+import talkgpt.storage.Storage;
+import talkgpt.task.*;
+
+public class ListCommand extends Command {
+
+    public ListCommand() {}
+
+    @Override
+    public boolean execute(TaskList list, Storage storage, UI ui) {
+        list.listTasks(ui);
+        return false;
+    }
+}
