+++ src/main/java/talkgpt/task/ToDos.java
@@ -1,28 +1,30 @@
+package talkgpt.task;
+
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
-class ToDos extends Task{
+public class ToDos extends Task{
 
-    ToDos(int index, String description) {
+    public ToDos(int index, String description) {
         super(index, description);
     }
 
-    ToDos(int index, String description, boolean status) {
+    public ToDos(int index, String description, boolean status) {
         super(index, description, status);
     }
 
     @Override
-    LocalDateTime getStart() {
+    public LocalDateTime getStart() {
         return LocalDateTime.now();
     }
 
     @Override
-    LocalDateTime getEnd() {
+    public LocalDateTime getEnd() {
         return LocalDateTime.now();
     }
 
     @Override
-    Task toggleStatus() {
+    public Task toggleStatus() {
         boolean newStatus = !super.getStatus();
         if (newStatus){
             System.out.println("Good Job on completing your task! I've marked this task!");
@@ -34,7 +36,7 @@ class ToDos extends Task{
     }
 
     @Override
-    boolean isValid() {
+    public boolean isValid() {
         if (super.getDescription().isEmpty()) {
             System.out.println("Please enter a description.");
             return false;
@@ -44,12 +46,12 @@ class ToDos extends Task{
     }
 
     @Override
-    boolean isDueOn(LocalDate dueDate) {
+    public boolean isDueOn(LocalDate dueDate) {
         return false;
     }
 
     @Override
-    String toFileFormat() { // D | id | isDone | description
+    public String toFileFormat() { // D | id | isDone | description
         return "T | " + super.getId() + " | " + (super.getStatus() ? "1" : "0") + " | " + super.getDescription();
     }
